requirements.txt
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
langchain==0.3.27
langchain-core==0.3.75
langchain-community==0.3.29
langchain-google-genai==2.0.10
langchain_text_splitters==0.3.11
pymupdf==1.26.4
python-dotenv==1.1.1
faiss-cpu==1.12.0